TST: shift dataset tests
